backend/requirements.txt
aioboto3==13.4.0
aiobotocore==2.18.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioitertools==0.12.0
aiosignal==1.3.2
amqp==5.3.1
aniso8601==10.0.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
boto3==1.36.1
botocore==1.36.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
Flask-Argon2==0.3.0.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
frozenlist==1.5.0
git-filter-repo==2.47.0
greenlet==3.1.1
gunicorn==23.0.0
hvac==2.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector-python==9.1.0
packaging==24.2
prompt_toolkit==3.0.50
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
